Added 3 new proejcts and updated the main project
I have changed the main project, added a new class for position, this contains the row and the column of a product. And I updated the UML Diagram. I think that, the relationships between all classes are aggregation (example: a ProductBand can exists without a ContainableItem, that means the band is empty), but I'm not sure that is correct.
The 3 new projects that I have implemented are: UnitTests, ConsoleUI and FormUI.
This commit doesn't include the implementation for Dispenser.